refactor: 문자열 계산기 클래스 분리
